requirements.txt
pandas>=2.2.2
numpy>=1.26.0
scikit-learn>=1.4.0
PyYAML>=6.0.1
pytorch-lightning>=2.2.0
pytorch-forecasting>=1.0.0
torch>=2.2.0
torchmetrics>=1.3.0
matplotlib>=3.8.2
seaborn>=0.13.0
rich>=13.7.0
joblib>=1.3.0
